+++ src/tests/integration/components/space-automation/input-stores-form-test.js
@@ -129,7 +129,7 @@ const storeTypes = {
     allowedDataSpecConfigs: dataSpecConfigsArray,
     dataSpecConfigKey: 'itemDataSpec',
     editors: dataSpecSpecificEditors,
-    complexValuesGenerator: arrayLikeValuesGenerator,
+    complexContentsGenerator: arrayLikeValuesGenerator,
   },
   // TODO: VFS-7816 uncomment or remove future code
   // map: {
@@ -137,7 +137,7 @@ const storeTypes = {
   //   editors: {
   //     default: 'rawValue',
   //   },
-  //   complexValuesGenerator: simpleValues =>
+  //   complexContentsGenerator: simpleValues =>
   //     simpleValues.map(simpleValue => `{"k0":${simpleValue}}`)
   //     .concat(simpleValues.length > 1 ? [
   //       `{${simpleValues.map((simpleValue, i) => `"k${i}":${simpleValue}`).join(',')}}`,
@@ -155,13 +155,13 @@ const storeTypes = {
     editors: {
       default: 'filesValue',
     },
-    complexValuesGenerator: () => [],
+    complexContentsGenerator: () => [],
   },
   singleValue: {
     allowedDataSpecConfigs: dataSpecConfigsArray,
     dataSpecConfigKey: 'itemDataSpec',
     editors: dataSpecSpecificEditors,
-    complexValuesGenerator: simpleValues => simpleValues,
+    complexContentsGenerator: simpleValues => simpleValues,
     filesLimit: 1,
   },
   // TODO: VFS-7816 uncomment or remove future code
@@ -172,12 +172,12 @@ const storeTypes = {
   //   editors: {
   //     default: 'rawValue',
   //   },
-  //   complexValuesGenerator: arrayLikeValuesGenerator,
+  //   complexContentsGenerator: arrayLikeValuesGenerator,
   // },
   // auditLog: {
   //   allowedDataSpecConfigs: dataSpecConfigsArray,
   //   editors: dataSpecSpecificEditors,
-  //   complexValuesGenerator: arrayLikeValuesGenerator,
+  //   complexContentsGenerator: arrayLikeValuesGenerator,
   // },
 };
 
@@ -287,7 +287,7 @@ describe('Integration | Component | space automation/input stores form', functio
     allowedDataSpecConfigs,
     dataSpecConfigKey,
     editors,
-    complexValuesGenerator,
+    complexContentsGenerator,
   }) => {
     allowedDataSpecConfigs.forEach(({
       name: dataSpecName,
@@ -320,37 +320,37 @@ describe('Integration | Component | space automation/input stores form', functio
         });
 
         if (editor === 'rawValue') {
-          const correctInitialValues = complexValuesGenerator(correctValues);
-          const incorrectInitialValues = complexValuesGenerator(incorrectValues);
+          const correctInitialContents = complexContentsGenerator(correctValues);
+          const incorrectInitialContents = complexContentsGenerator(incorrectValues);
 
-          correctInitialValues.forEach(initialValue => {
-            it(`recognizes ${initialValue} value as valid`, async function () {
+          correctInitialContents.forEach(initialContent => {
+            it(`recognizes ${initialContent} value as valid`, async function () {
               const changeSpy = this.get('changeSpy');
               await render(this);
 
-              await fillIn(`.${editor}-field .form-control`, initialValue);
+              await fillIn(`.${editor}-field .form-control`, initialContent);
 
               expect(this.$(`.${editor}-field`)).to.not.have.class('.has-error');
               expect(changeSpy).to.be.calledWith({
                 data: {
-                  s1: JSON.parse(initialValue),
+                  s1: JSON.parse(initialContent),
                 },
                 isValid: true,
               });
             });
           });
 
-          incorrectInitialValues.forEach(initialValue => {
-            it(`recognizes ${initialValue} value as invalid`, async function () {
+          incorrectInitialContents.forEach(initialContent => {
+            it(`recognizes ${initialContent} value as invalid`, async function () {
               const changeSpy = this.get('changeSpy');
               await render(this);
 
-              await fillIn(`.${editor}-field .form-control`, initialValue);
+              await fillIn(`.${editor}-field .form-control`, initialContent);
 
               expect(this.$(`.${editor}-field`)).to.have.class('has-error');
               expect(changeSpy).to.be.calledWith({
                 data: {
-                  s1: JSON.parse(initialValue),
+                  s1: JSON.parse(initialContent),
                 },
                 isValid: false,
               });
@@ -358,7 +358,7 @@ describe('Integration | Component | space automation/input stores form', functio
           });
 
           it('fills initial value with JSON', async function () {
-            const rawValue = JSON.parse(correctInitialValues[0]);
+            const rawValue = JSON.parse(correctInitialContents[0]);
             getStores(this)[0].defaultInitialContent = rawValue;
 
             await render(this);